Add support for asynchronous file I/O on Linux.
This patch adds asynchronous file reads to epoll event loops via Linux asynchronous I/O and
libaio. The bulk of the code is implemented in AIOContext, Java side, and in netty_epoll_native.c,
native side.

When AIO is available and the user requests it, the AIOContext is added to an epoll event loop,
and the AIO epoll descriptor is monitored along with other epoll notifications. The AIOContext
methods should only be run by the thread of its owning event loop. AIOContext has no locking or
any thread safety, since it doesn't need to.

With this patch, users can then read files asynchronously without any other thread required but
by relying directly on the epoll event loop. To do so they create an asynchrounous file channel,
currently called AIOEpollFileChannel, and register it with an epoll event loop.

Limitations:

- Only reads are currently supported.

- Libaio must be installed and if it is missing then the entire epoll netty library will not load.
This is obviously not ideal and should be addressed before merging. We could either separate
out the aio native code in its own library or, with a bit of extra work, we could probably
remove libaio altogether since it's only a thin wrapper of the OS system calls.

- AIOEpollFileChannel accepts an event loop in the contructor and directly in the read calls.
Passing it to the constructor is more elegant, but in practice one wants to read a file without
switching to another event loop thread, and typically the thread creating a file channel may not
be the same as the threads reading it. This is why we then ended up adding the EL to the read
methods as well. So a file channel could be read by multiple AIO contexts at the same time.

- This patch also contains a couple of unrelated changes that allow extending Nio and Epoll event
loops, these changes should be moved to a separate pull request.